(refactor): oracle
I refactored the encryption oracle to have a bit more flexibility in its
usage. The refactor also allows for reuse of the same key generated. I
defined a "light" Oracle interface (just a closure type really) and changed the
oracle function to be a constructor that returns an Oracle closure that
can then be called by specifying the desired mode. Due to this now being
a constructor, it can be called multiple times while using the same key.
The changes should allow for more flexibility and easier testing.