(refactor): cryptanalysis
Changed the cryptanalysis package name to analysis for brevity's sake. I
also updated a variety of functions with underscores to remove the
underscores as they are not in line with Go's function naming best
practices. It's not noted above, but I also implemented the blocksize
detection algorithm in the analysis package.