(feat): Challenge 13 scaffolding
Added the profile generation and conversion functions to the parser.go
file, which I put in the encoding package. Not a perfect fit but seems
to make sense for now - may move it later. There are TODOs in the
Challenge13() function outlining some basic next steps for the ECB block
substitution attack.