(refactor): analysis
Added a function in the analysis package FindRepeatingBlocks, which
detects the index for the first of one or more contiguous equal blocks.
This will be used in the FindAlignment function to determine the number
of prepended random bytes for a given oracle. This will allow me to
determine the appropriate offset in the ciphertext to pass to my byte at
a time decryption function.