(refactor): FindAlignment
Implemented the FindAlignment function in the analysis package to
determine the padding length of randomly prepended bytes. This is
accomplished by adding an increasing number of bytes, then comparing the
two following blocks for equivalence by supplying the 2 blocks worth of
the same character. I also fixed a bug in the FindRepeatingBlocks
function to ensure we don't slice past the end of the data.