(refactor): finished challenge 14
Added a WrapOracle function to wrap the existing oracle to align the
oracle for the random prepended bytes. I had to fix a bug in my
NewOracle function as I had included the generation of an optional
prefix and suffix for a given oracle in the closure itself and not prior
to the closure. The result was that each time the oracle was called it
generated a new random prefix or suffix. This obviously caused issued
with Challenge 14 and needed to be fixed. Now, both are generated prior
to the closure and used within the closure if defined.