(feat): challenge 16
Completed challenge 16. Updated the buildCiphertext function to be a
buidInput function that returns a full input string given a
user-controlled string. I updated the oracle to allow for providing a
key and iv such that decryption can be accomplished outside of the
oracle, which was necessary for this challenge. I also moved padding to
occur inside of the oracle as opposed to outside the oracle. I also
implemented the bitflipping attack in a new file in the analysis
package: cbc.go. This takes an oracle, desired plaintext, and a string
builder style function (buildInput in this case). It will build the
input, encrypt it, identify the control block and bit flip to get the
desired ciphertext.